chore: experiments with agents
