+++ agents/agents_interface.go
@@ -2,12 +2,13 @@ package agents
 
 import (
 	"context"
-	//"github.com/TobiasGleiter/langchain-go/tools"
+	
+	"github.com/TobiasGleiter/langchain-go/tools"
 )
 
 type Agent interface {
 	Plan(ctx context.Context, steps []AgentStep, inputs map[string]string) ([]AgentAction, *AgentFinish, error)
-	GetTools() []Tool
+	GetTools() []tools.Tool
 	GetInputKeys() []string
 	GetOutputKeys() []string
 }
