+++ agents/agents_types.go
@@ -1,18 +1,15 @@
 package agents
 
-// type AgentAction struct {
-// 	Tool      string
-// 	ToolInput string
-// 	Log       string
-// 	ToolID    string
-// }
+type AgentAction struct {
+	Tool      string
+	ToolInput string
+}
 
-// type AgentStep struct {
-// 	Action      AgentAction
-// 	Observation string
-// }
+type AgentStep struct {
+	Action      AgentAction
+	Observation string
+}
 
-// type AgentFinish struct {
-// 	ReturnValues map[string]any
-// 	Log          string
-// }
+type AgentFinish struct {
+	ReturnValues map[string]any
+}
