+++ tools/tool_interface.go
@@ -3,6 +3,6 @@ package tools
 import ("context")
 
 type Tool interface {
-    Name() string
-    Call(ctx context.Context, input string) (string, error)
+	Name() string
+	Call(ctx context.Context, input string) (string, error)
 }
